fix infer by adding pyssam and use stpyvista to render
