+++ utils/run_exodus_to_html_scene.py
@@ -3,6 +3,39 @@ from os import makedirs
 from sys import argv
 import pyvista as pv
 
+def show_geom(inputfile, rendering, show_vacuum=False):
+    plotter = pv.Plotter(window_size=[400,400])
+
+    # Create a mesh with a cube 
+    if type(inputfile) == str:
+        mesh = pv.read(inputfile)
+    else:
+        print("input file is pyvista object")
+        mesh = inputfile
+    if rendering.lower() == "metal":
+        # plot coil with RGB for copper
+        plotter.add_mesh(mesh.get(0)["coil"], color=[184, 115, 51])
+        # plot target with RGB for copper
+        plotter.add_mesh(mesh.get(0)["target"], color=[225,229,233])
+        if show_vacuum:
+            plotter.add_mesh(mesh.get(0)["vacuum_region"], opacity=0.2)
+    elif rendering.lower() == "field":
+        cmap = "plasma"
+        plotter.add_mesh(mesh, cmap=cmap)
+    elif rendering.lower() == "none":
+        for block_name in mesh.get(0).keys():
+            # check for vacuum, and avoid showing it
+            if "vacuum" in block_name.lower():
+                if show_vacuum:
+                    plotter.add_mesh(mesh.get(0)[block_name], opacity=0.2)
+            else:
+                plotter.add_mesh(mesh.get(0)[block_name])
+
+    # Final touches
+    plotter.view_isometric()
+    pv.global_theme.transparent_background = True
+    return plotter
+
 
 def get_inputs():
     parser = argparse.ArgumentParser()
@@ -37,33 +70,13 @@ def get_inputs():
     )
     return parser.parse_args()
 
-args = get_inputs()
 
-plotter = pv.Plotter(window_size=[400,400])
+if __name__ == "__main__":
 
-# Create a mesh with a cube 
-mesh = pv.read(args.inputfile)
-if args.rendering.lower() == "metal":
-    # plot coil with RGB for copper
-    plotter.add_mesh(mesh.get(0)["coil"], color=[184, 115, 51])
-    # plot target with RGB for copper
-    plotter.add_mesh(mesh.get(0)["target"], color=[225,229,233])
-    if args.show_vacuum:
-        plotter.add_mesh(mesh.get(0)["vacuum_region"], opacity=0.2)
-elif args.rendering.lower() == "field":
-    cmap = "plasma"
-    plotter.add_mesh(mesh, cmap=cmap)
-elif args.rendering.lower() == "none":
-    for block_name in mesh.get(0).keys():
-        # check for vacuum, and avoid showing it
-        if "vacuum" in block_name.lower():
-            if args.show_vacuum:
-                plotter.add_mesh(mesh.get(0)[block_name], opacity=0.2)
-        else:
-            plotter.add_mesh(mesh.get(0)[block_name])
+    plotter = pv.Plotter(window_size=[400,400])
 
-# Final touches
-plotter.view_isometric()
-pv.global_theme.transparent_background = True
-makedirs("tmp_data", exist_ok=True)
-plotter.export_html(f'tmp_data/{args.outputfile}')
+    args = get_inputs()
+    plotter = show_geom(args.inputfile, args.rendering, args.show_vacuum)
+    pv.global_theme.transparent_background = True
+    makedirs("tmp_data", exist_ok=True)
+    plotter.export_html(f'tmp_data/{args.outputfile}')
